Add files via upload
Used the new and improved data pipeline to clean the QnA_Biomass_f1000_20230717 document which contains 10 000+ QnAs. The pipeline currently does the following corrections from the QnA data issues list : 
Compound words without hyphen
Compound words with hyphen
Proper noun
Grammatical errors
adds space between brackets and words
Incorrect capital
multiple comma clusters
URL errors
non formatted numbers errors
extra character at the beginning of the sentences
no uppercase in initial word.

It also filters the truncated questions and answers in the document fed to it
and does away with them and then stores the undesired questions in a document with the suffix 'filtered_questions'.
It then stores the cleaned QnA in the downloads directory, in a file called cleaned_QnAs_non_optimized (works for windows file directory only). 


Things to note:
a function called def identify_and_capitalize_acronyms is contained in the code however not called in the execution of the 'main' function (it was actually called but then commented out). 
This is because the function doesn't work well such that, it capitalizes acronyms that are not English stop-words and that are not names of persons but are other parts of speech. In other words, the code would be unable to tell apart parts of speech from actual acronyms which is a problem. This should be looked into.

Furthermore, the part of the code that counts the errors has been commented out to increase the speed of the code thereby reducing the runtime. These comments exist in the definition of the LanguageTool function and the calling of the same function in the 'main' function.